requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic-settings
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
python-multipart
